requirements.txt
arrow==0.13.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2
geographiclib==1.49
geopy==1.19.0
idna==2.8
instaloader==4.2.4
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
units==0.7
urllib3==1.24.2
